+++ src/cupertino-pane.ts
@@ -10,8 +10,7 @@ export class CupertinoPane {
     parentElement: null,
     followerElement: null,
     backdrop: false,
-    backdropOpacity: 0.4, 
-    animationType: 'ease',
+    backdropOpacity: 0.4,
     animationDuration: 300,
     dragBy: null,
     bottomOffset: 0,
@@ -43,10 +42,11 @@ export class CupertinoPane {
   };
 
   private defaultBreaksConf = {
-    top: { enabled: true, height: window.innerHeight - (135 * 0.35)},
-    middle: { enabled: true, height: 300},
-    bottom: { enabled: true, height: 100},
+    top: { enabled: true, height: window.innerHeight - (135 * 0.35), timing: 'ease'},
+    middle: { enabled: true, height: 300, timing: 'cubic-bezier(0.175, 0.885, 0.370, 1.120)'},
+    bottom: { enabled: true, height: 100, timing: 'cubic-bezier(0.175, 0.885, 0.370, 1.120)'},
   };
+  
   private screen_height: number = window.innerHeight;
   private steps: any[] = [];
   private startP: any;
@@ -167,7 +167,7 @@ export class CupertinoPane {
       // Content
       this.contentEl = this.el;
       this.contentEl.style.display = 'block';
-      this.contentEl.style.transition = `opacity ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+      this.contentEl.style.transition = `opacity ${this.settings.animationDuration}ms ease 0s`;
       this.contentEl.style.overflowX = 'hidden';
 
       // Close button
@@ -216,6 +216,13 @@ export class CupertinoPane {
           this.settings.breaks[val] = this.defaultBreaksConf[val];
         }
   
+        // Set timings by default 
+        if (this.settings.breaks[val]
+          && this.settings.breaks[val].enabled
+          && !this.settings.breaks[val].timing) {
+            this.settings.breaks[val].timing = this.defaultBreaksConf[val].timing;
+        }
+
         // Add offsets (offset or height, later need remove ofsfset)
         if (this.settings.breaks[val]
             && this.settings.breaks[val].enabled
@@ -273,7 +280,7 @@ export class CupertinoPane {
         );
         this.followerEl.style.willChange = 'transform, border-radius';
         this.followerEl.style.transform = `translateY(0px) translateZ(0px)`;
-        this.followerEl.style.transition = `all ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+        this.followerEl.style.transition = `all ${this.settings.animationDuration}ms ${this.settings.breaks[this.currentBreak()].timing} 0s`;
       }
 
       if (!this.settings.showDraggable) {
@@ -406,7 +413,7 @@ export class CupertinoPane {
     let attrElements = this.el.querySelectorAll('[hide-on-bottom]');
     if (!attrElements.length) return;
     attrElements.forEach((item) => {
-      (<HTMLElement>item).style.transition = `opacity ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+      (<HTMLElement>item).style.transition = `opacity ${this.settings.animationDuration}ms ease 0s`;
       (<HTMLElement>item).style.opacity = (val >= this.breaks['bottom']) ? '0' : '1';
     });
   }
@@ -649,7 +656,7 @@ export class CupertinoPane {
     this.backdropEl.style.right = '0';
     this.backdropEl.style.left = '0';
     this.backdropEl.style.top = '0';
-    this.backdropEl.style.transition = `all ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+    this.backdropEl.style.transition = `all ${this.settings.animationDuration}ms ease 0s`;
     this.backdropEl.style.backgroundColor = `rgba(0,0,0, ${this.settings.backdropOpacity})`;
     this.backdropEl.style.display = 'none';
     this.backdropEl.style.zIndex = '10';
@@ -678,7 +685,7 @@ export class CupertinoPane {
       this.backdropEl.removeEventListener('transitionend', transitionEnd); 
     }
     
-    this.backdropEl.style.transition = `all ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+    this.backdropEl.style.transition = `all ${this.settings.animationDuration}ms ease 0s`;
     this.backdropEl.style.backgroundColor = 'rgba(0,0,0,.0)';
 
     if (!conf.show) {
@@ -956,7 +963,7 @@ export class CupertinoPane {
             || params.type === 'destroy'
             || params.type === 'present') {
           this.backdropEl.style.backgroundColor = 'rgba(0,0,0,.0)';
-          this.backdropEl.style.transition = `all ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+          this.backdropEl.style.transition = `all ${this.settings.animationDuration}ms ease 0s`;
           
           if (params.type !== 'hide' && params.type !== 'destroy') {
             this.backdropEl.style.display = 'block';
@@ -969,12 +976,19 @@ export class CupertinoPane {
       
       // freemode
       if (params.type === 'end' && this.settings.freeMode) return; 
+      
+      // Get timing function for next break
+      // TODO: getBreakByHeight()
+      const nextBreak = Object.entries(this.settings.breaks).find(
+        val => val[1].height === (this.screen_height - params.translateY)
+      );
+      const timingForNext = nextBreak ? nextBreak[1].timing : 'ease';
 
       // style
-      this.paneEl.style.transition = `transform ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+      this.paneEl.style.transition = `transform ${this.settings.animationDuration}ms ${timingForNext} 0s`;
       // Bind for follower same transitions
       if (this.followerEl) {
-        this.followerEl.style.transition = `transform ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+        this.followerEl.style.transition = `transform ${this.settings.animationDuration}ms ${timingForNext} 0s`;
       }
 
       // Main transitions
@@ -982,9 +996,9 @@ export class CupertinoPane {
         this.paneEl.style.transform = `translateY(${this.screen_height}px) translateZ(0px)`;
         setTimeout(() => {
           // Emit event
-          this.settings.onTransitionStart({translateY: {new: this.breaks[this.settings.initialBreak]}});
+          this.settings.onTransitionStart({translateY: {new: params.translateY}});
 
-          this.paneEl.style.transform = `translateY(${this.breaks[this.settings.initialBreak]}px) translateZ(0px)`;
+          this.paneEl.style.transform = `translateY(${params.translateY}px) translateZ(0px)`;
           // Bind for follower same transitions
           if (this.followerEl) {
             this.followerEl.style.transform = `translateY(0px) translateZ(0px)`;