+++ src/models.ts
@@ -1,7 +1,7 @@
 export interface PaneBreaks {
-  top?: {enabled: boolean, height?: number, offset?: number};
-  middle?: {enabled: boolean, height?: number, offset?: number};
-  bottom?: {enabled: boolean, height?: number, offset?: number};
+  top?: {enabled: boolean, height?: number, offset?: number, timing?: string};
+  middle?: {enabled: boolean, height?: number, offset?: number, timing?: string};
+  bottom?: {enabled: boolean, height?: number, offset?: number, timing?: string};
 }
 
 export interface TransitionStartEvent {
@@ -14,7 +14,6 @@ export interface Settings {
   followerElement: string;
   backdrop: boolean;
   backdropOpacity: number;
-  animationType: ('ease'|'ease-in'|'ease-out'|'ease-in-out'|'cubic-bezier');
   animationDuration: number;
   bottomOffset: number,
   darkMode: boolean;